Enable writing via DoPut
